dm-light/requirements.txt
requests~=2.31.0
dnspython~=2.6.1
ipaddress~=1.0.23
configparser~=6.0.1

asyncssh~=2.15.0
dnslib~=0.9.25